Some cleanup and getting submit-on-enter to work better
