+++ src/Eviden.VirtualGrocer/Server/Eviden.VirtualGrocer.Web.Server.csproj
@@ -46,7 +46,7 @@
     <PackageReference Include="Microsoft.Identity.Web.UI" Version="2.13.2" />
     <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.Server" Version="7.0.5" />
     <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="7.0.0" />
-    <PackageReference Include="Microsoft.SemanticKernel" Version="0.14.547.1-preview" />
+    <PackageReference Include="Microsoft.SemanticKernel" Version="0.19.230804.2-preview" />
   </ItemGroup>
 
   <ItemGroup>
