+++ src/Eviden.VirtualGrocer/Server/Skills/History/ITokenCounter.cs
@@ -1,0 +1,7 @@
+﻿namespace Eviden.VirtualGrocer.Web.Server.Skills.History
+{
+    public interface ITokenCounter
+    {
+        int CountTokens(string message);
+    }
+}
