+++ src/Eviden.VirtualGrocer/Server/Skills/History/TokenCounter.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.SemanticKernel.Connectors.AI.OpenAI.Tokenizers;
+
+namespace Eviden.VirtualGrocer.Web.Server.Skills.History
+{
+    public class TokenCounter : ITokenCounter
+    {
+        public int CountTokens(string message) => GPT3Tokenizer.Encode(message).Count;
+    }
+}
