+++ src/Eviden.VirtualGrocer/Server/Skills/InventorySearchSkill.cs
@@ -3,6 +3,7 @@ using Eviden.VirtualGrocer.Shared.Models;
 using Eviden.VirtualGrocer.Web.Server.Models;
 using Microsoft.SemanticKernel.Orchestration;
 using Microsoft.SemanticKernel.SkillDefinition;
+using System.ComponentModel;
 using System.Text.Json;
 
 namespace Eviden.VirtualGrocer.Web.Server.Skills
@@ -14,27 +15,35 @@ namespace Eviden.VirtualGrocer.Web.Server.Skills
         public InventorySearchSkill(SearchClient searchClient) =>
             _searchClient = searchClient;
 
-        [SKFunction("Search inventory")]
-        [SKFunctionName(SkillNames.FindInventory)]
+        [SKFunction]
+        [SKName(SkillNames.FindInventory)]
+        [Description("Search inventory")]
         public async Task<string> FindInventory(string query, SKContext context)
         {
             if (string.IsNullOrEmpty(query))
             {
-                context["products"] = "[]";
+                context.Variables["products"] = "[]";
                 return string.Empty;
             }
 
             var response = await _searchClient.SearchAsync<ProductSearchResult>(query);
-            var results = (await response.Value.GetResultsAsync().ToListAsync()).Select(x => x.Document);
+            var results = response.Value.GetResultsAsync();
 
-            string products = JsonSerializer.Serialize(results);
-            context["products"] = products;
-            return products;
+            List<ProductSearchResult> products = new List<ProductSearchResult>();
+            await foreach (var result in results)
+            {
+                products.Add(result.Document);
+            }
+
+            string productString = JsonSerializer.Serialize(products);
+            context.Variables["products"] = productString;
+            return productString;
         }
 
-        [SKFunction("Build a search query from the input.")]
-        [SKFunctionContextParameter(Name = "shoppingList")]
-        [SKFunctionName(SkillNames.BuildInventoryQuery)]
+        [SKFunction]
+        [SKName(SkillNames.BuildInventoryQuery)]
+        [Description("Build a search query from the input.")]
+        [SKParameter("shoppingList", "The shopping list result")]
         public string BuildInventoryQuery(SKContext context)
         {
             var itemJson = context.Variables["shoppingList"];
